% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bob_quote.R
\name{bob_quote}
\alias{bob_quote}
\title{Generate a random Bob Ross quote}
\usage{
bob_quote()
}
\description{
This function allows you to generate a random Bob Ross quote
}
\examples{
bob_quote()
}
\keyword{quote}
